requirements.txt
requests
streamlit
beautifulsoup4
fonttools
brotli
pyinstaller
browser_cookie3
websocket-client
